egondosi - added IEFA.py as single run script that also queries user for inputs
